feat: Create train_classifier script to build and train NLP classifier
The script takes the following inputs:
- database_filepath: Location of SQL database storing disaster messages
- table_name (optional): Name of table in database_filepath if it
  is named something other than 'merged'
- model_filepath: Location to store trained model as a pickle file

The train_classifier script loads the database, builds a machine
learning pipeline using sklearn's Pipeline object and custom
tokenizing functions on both the words and parts of speech. Then it
performs multioutput classification to train the model on 35 of the
original 36 disaster response categories (excluding 'child_alone'
since there was no messages to train that category on).  Finally,
the script performs a grid search to optimize the model hyperparameters
before predicting and scoring the performance of the model,
displaying the performance results, and saving the best model to a
pickle file.